requirements.txt
fastapi
uvicorn
python-docx
youtube-transcript-api
google-api-python-client
python-multipart
